Revert "Feat/방생성 복구 (#37)"
This reverts commit 83ba1282883a97a4950ec3cde04b7eb91d47503c.
