+++ src/main/java/com/example/web2_3_ourtuft_be/room/entity/Room.java
@@ -2,7 +2,6 @@ package com.example.web2_3_ourtuft_be.room.entity;
 
 import com.example.web2_3_ourtuft_be.item.game.entity.Game;
 import com.example.web2_3_ourtuft_be.quiz.entity.QuizSet;
-
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -12,21 +11,25 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 public class Room {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "room_id", nullable = false)
     private Long id;
+
     @ManyToOne
     @JoinColumn(name = "game_id")
     private Game gameId;
+
     @ManyToOne
     @JoinColumn(name = "quiz_id")
     private QuizSet quizId;
+
     @Column(nullable = false)
     private String roomName;
 
     private boolean disclosure;
-    private String roomPassword;
+    private int roomPassword;
     private int peopleEntering;
     private int round;
     private String host;
